+++ Scripts/package_app.sh
@@ -110,7 +110,7 @@ cat > "$CONTENTS_DIR/Info.plist" <<EOF
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleShortVersionString</key>
-    <string>0.8.10</string>
+    <string>0.8.11</string>
     <key>CFBundleVersion</key>
     <string>1</string>
     <key>CFBundleLocalizations</key>